Refactored the project structure to better contain the godot things.
