+++ godot/README.md
@@ -1,0 +1,3 @@
+# Godot dedicated directory
+This directory contains the Godot related `Component`s, `System`s, `Resource`s, `Node`s, `Editor Plugin`s.
+
