+++ godot/editor_plugins/entity_editor_plugin.cpp
