+++ godot/nodes/ecs_utilities.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../ecs.h"
+#include "../../ecs.h"
 #include "core/io/resource.h"
 #include "core/templates/local_vector.h"
 #include "core/templates/oa_hash_map.h"
