+++ tests/test_ecs_query.h
@@ -4,8 +4,8 @@
 #include "tests/test_macros.h"
 
 #include "../components/dynamic_component.h"
-#include "../components/transform_component.h"
 #include "../ecs.h"
+#include "../godot/components/transform_component.h"
 #include "../iterators/dynamic_query.h"
 #include "../world/world.h"
